requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
feedparser>=6.0.11
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.1.0
apscheduler>=3.10.4
werkzeug>=3.0.1
openai>=1.30.0
python-dateutil>=2.9.0
pytz>=2024.1
